requirements.txt
loguru>=0.7.0
pandas>=2.1.0
PyYAML>=6.0
ccxt>=4.2.0
openpyxl>=3.1.0
